Enforce checking and updating license and copying file
The copyright dates where not correctly updated and were also out of
sync in some cases. If a valid copyright date range is detected in a
source code file, both files are now checked and updated if needed.
Also before one of the files is updated, the other file is checked as
well to calculate the correct date range and make sure they are in sync.