Added *experimental* sniff WhiteSpace.ScopeIndent sniff
It implements stricter rules and replaces two other sniffs:
- Generic.WhiteSpace.ScopeIndent
- PSR2.Methods.FunctionCallSignature.Indent
- Squiz.WhiteSpace.ScopeClosingBrace.Indent

It requires more tests to find out if all cases are there already included
and if it works correctly with multiple open/close php tags.

This sniff is not working correctly in all cases, because of bug:
https://github.com/squizlabs/PHP_CodeSniffer/pull/1532
It will be fixed in PHP_CodeSniffer 3.0.2.